+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - **Breaking:** envs: Set rewards of all environments to zero, left to be decided by agents
+- Makefile: Remove `UPKIE_NAME` env variable, we now expect an `upkie` host in the SSH config
 
 ### Removed